requirements.txt
django==3.1.7
djangorestframework==3.12.2
django-filter==2.4.0
markdown==3.3.4
python-decouple==3.4
dj-database-url==0.5.0